Move database representation into sandbox module.
