+++ src/main/java/edu/kit/datamanager/pit/pidsystem/impl/sandbox/MinPidRecordDao.java
@@ -1,4 +1,4 @@
-package edu.kit.datamanager.pit.domain.minimal;
+package edu.kit.datamanager.pit.pidsystem.impl.sandbox;
 
 import java.util.Optional;
